# [server-side apply] Cap the number of managedFields entries for updates at 10