… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/capmanagers_test.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/fieldmanager_test.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/managedfields_test.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/skipnonapplied.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/stripmeta.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/structuredmerge.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/patch.go
… staging/src/k8s.io/apiserver/pkg/endpoints/installer.go
… test/integration/apiserver/apply/apply_test.go