**What type of PR is this?**
/kind feature
/sig api-machinery
/wg apply
/priority important-soon
/cc @apelisse @lavalamp @kwiesmueller 

**What this PR does / why we need it**:
Prevents the list of managers from growing too large.

also,
Fixes #82042

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```