@apelisse 2019-09-11
I saw this PR while searching for your "split fieldmanager" PR. Can we do sort-of the same thing for this? Build it as part of a separate wrapper class?