@apelisse 2019-09-11
> I saw this PR while searching for your "split fieldmanager" PR. Can we do sort-of the same thing for this? Build it as part of a separate wrapper class?

We might have to parse the managers again. Maybe we should come-up with an alternative solution for that (lazy-deserialization of the manager fields with some sort of caching)