@lavalamp 2019-09-20
I would not expect the cap to apply to appliers.

On Fri, Sep 20, 2019 at 12:41 PM Jennifer Buckley <notifications@github.com>
wrote:

> *@jennybuckley* commented on this pull request.
> ------------------------------
>
> In
> staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/fieldmanager.go
> <https://github.com/kubernetes/kubernetes/pull/81816#discussion_r326775414>
> :
>
> > +		// Create a new manager identifier for the versioned bucket entry.
> +		// The version for this manager comes from the version of the update being merged into the bucket.
> +		if manager, err = internal.BuildManagerIdentifier(&metav1.ManagedFieldsEntry{
> +			Manager:    "manager-too-old-to-track",
> +			Operation:  metav1.ManagedFieldsOperationUpdate,
> +			APIVersion: string(vs.APIVersion()),
> +		}); err != nil {
> +			return managed, fmt.Errorf("failed to create bucket manager for version %v: %v", string(vs.APIVersion()), err)
> +		}
> +
> +		// Merge the fieldets if the bucket already exists, otherwise just create a new entry.
> +		// Use the time from the update being merged into the bucket, since it is more recent.
> +		if previous, ok := managed.Fields[manager]; ok {
> +			managed.Fields[manager] = fieldpath.NewVersionedSet(vs.Set().Union(previous.Set()), vs.APIVersion(), vs.Applied())
> +		} else {
> +			managed.Fields[manager] = vs
>
> but then what if you have more than 10 appliers?
>
> Probably the same thing as it would do if you had more than 10
> apiversions, just keep the number of updaters to a minimum
>
> And what if you have 8 appliers, does it mean you have at most 2
> non-appliers?
>
> Yes
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/81816?email_source=notifications&email_token=AAE6BFS37LY36WA2VXUF34TQKURP5A5CNFSM4IO2NIM2YY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCFO426I#discussion_r326775414>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFT4BDJALWM7WPKFOXDQKURP5ANCNFSM4IO2NIMQ>
> .
>