@jennybuckley 2019-09-23
```
jennybuckley@jennybuckley:~/Projects/k8s-p3/src/k8s.io/kubernetes$ benchcmp ~/old.txt ~/new.txt
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkNoServerSideApply/List1-12                          1137225       1127525       -0.85%
BenchmarkNoServerSideApply/List20-12                         2217764       2202924       -0.67%
BenchmarkNoServerSideApply/List200-12                        12631556      12876137      +1.94%
BenchmarkNoServerSideApply/List2000-12                       78666192      78537239      -0.16%
BenchmarkNoServerSideApply/RepeatedUpdates-12                3023928       3013982       -0.33%
BenchmarkNoServerSideApply/Post1-12                          4080269       4029934       -1.23%
BenchmarkNoServerSideApply/Post10-12                         11044128      8094545       -26.71%
BenchmarkNoServerSideApply/Post50-12                         23971097      23472944      -2.08%
BenchmarkNoServerSideApplyButSameSize/List1-12               1213408       1203910       -0.78%
BenchmarkNoServerSideApplyButSameSize/List20-12              2651249       2692173       +1.54%
BenchmarkNoServerSideApplyButSameSize/List200-12             18717204      18616164      -0.54%
BenchmarkNoServerSideApplyButSameSize/List2000-12            119203034     119825575     +0.52%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     4171596       3913753       -6.18%
BenchmarkNoServerSideApplyButSameSize/Post1-12               4806527       4868024       +1.28%
BenchmarkNoServerSideApplyButSameSize/Post10-12              9082722       8867710       -2.37%
BenchmarkNoServerSideApplyButSameSize/Post50-12              24727741      25535247      +3.27%
BenchmarkServerSideApply/List1-12                            1271745       1200407       -5.61%
BenchmarkServerSideApply/List20-12                           2857699       2859166       +0.05%
BenchmarkServerSideApply/List200-12                          20632891      20279522      -1.71%
BenchmarkServerSideApply/List2000-12                         129143255     132706920     +2.76%
BenchmarkServerSideApply/RepeatedUpdates-12                  4969736       4936952       -0.66%
BenchmarkServerSideApply/Post1-12                            6545111       6990540       +6.81%
BenchmarkServerSideApply/Post10-12                           11478310      11224251      -2.21%
BenchmarkServerSideApply/Post50-12                           38008987      37451440      -1.47%

benchmark                                                    old allocs     new allocs     delta
BenchmarkNoServerSideApply/List1-12                          603            603            +0.00%
BenchmarkNoServerSideApply/List20-12                         4729           4729           +0.00%
BenchmarkNoServerSideApply/List200-12                        43673          43676          +0.01%
BenchmarkNoServerSideApply/List2000-12                       432677         432669         -0.00%
BenchmarkNoServerSideApply/RepeatedUpdates-12                1931           1930           -0.05%
BenchmarkNoServerSideApply/Post1-12                          5390           5390           +0.00%
BenchmarkNoServerSideApply/Post10-12                         54344          54344          +0.00%
BenchmarkNoServerSideApply/Post50-12                         271996         271931         -0.02%
BenchmarkNoServerSideApplyButSameSize/List1-12               618            618            +0.00%
BenchmarkNoServerSideApplyButSameSize/List20-12              4933           4933           +0.00%
BenchmarkNoServerSideApplyButSameSize/List200-12             45699          45703          +0.01%
BenchmarkNoServerSideApplyButSameSize/List2000-12            452879         453013         +0.03%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     1985           1984           -0.05%
BenchmarkNoServerSideApplyButSameSize/Post1-12               5518           5519           +0.02%
BenchmarkNoServerSideApplyButSameSize/Post10-12              55606          55605          -0.00%
BenchmarkNoServerSideApplyButSameSize/Post50-12              278220         278158         -0.02%
BenchmarkServerSideApply/List1-12                            640            640            +0.00%
BenchmarkServerSideApply/List20-12                           5373           5373           +0.00%
BenchmarkServerSideApply/List200-12                          50128          50115          -0.03%
BenchmarkServerSideApply/List2000-12                         496920         496947         +0.01%
BenchmarkServerSideApply/RepeatedUpdates-12                  5223           5221           -0.04%
BenchmarkServerSideApply/Post1-12                            10359          10361          +0.02%
BenchmarkServerSideApply/Post10-12                           103985         104016         +0.03%
BenchmarkServerSideApply/Post50-12                           520067         520136         +0.01%

benchmark                                                    old bytes     new bytes     delta
BenchmarkNoServerSideApply/List1-12                          52523         52505         -0.03%
BenchmarkNoServerSideApply/List20-12                         540331        541573        +0.23%
BenchmarkNoServerSideApply/List200-12                        5187237       5195626       +0.16%
BenchmarkNoServerSideApply/List2000-12                       50044200      49985709      -0.12%
BenchmarkNoServerSideApply/RepeatedUpdates-12                153812        153852        +0.03%
BenchmarkNoServerSideApply/Post1-12                          585293        586859        +0.27%
BenchmarkNoServerSideApply/Post10-12                         5974824       5966362       -0.14%
BenchmarkNoServerSideApply/Post50-12                         29965720      29912661      -0.18%
BenchmarkNoServerSideApplyButSameSize/List1-12               71167         71168         +0.00%
BenchmarkNoServerSideApplyButSameSize/List20-12              919515        920376        +0.09%
BenchmarkNoServerSideApplyButSameSize/List200-12             9714920       9701023       -0.14%
BenchmarkNoServerSideApplyButSameSize/List2000-12            92360797      92194064      -0.18%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     211081        211198        +0.06%
BenchmarkNoServerSideApplyButSameSize/Post1-12               670251        670576        +0.05%
BenchmarkNoServerSideApplyButSameSize/Post10-12              6813230       6808404       -0.07%
BenchmarkNoServerSideApplyButSameSize/Post50-12              34104228      34079409      -0.07%
BenchmarkServerSideApply/List1-12                            72444         72500         +0.08%
BenchmarkServerSideApply/List20-12                           926285        926215        -0.01%
BenchmarkServerSideApply/List200-12                          9815583       9822102       +0.07%
BenchmarkServerSideApply/List2000-12                         94235371      93770744      -0.49%
BenchmarkServerSideApply/RepeatedUpdates-12                  513008        510596        -0.47%
BenchmarkServerSideApply/Post1-12                            1103706       1103615       -0.01%
BenchmarkServerSideApply/Post10-12                           11134184      11135302      +0.01%
BenchmarkServerSideApply/Post50-12                           55751166      55712693      -0.07%
```
Reran benchmarks to make sure this still doesn't make them worse, since the PR got more complicated